Add files via upload
Simple starter sketch for 12 segment bar graph. 
The code for register C had to be altered after I mistakenly soldered the second bar graph wires in reverse on the shift register. LEDs 2-8 are wired 8-2. If wiring the bar graphs correctly you can use the array for register A in place of the attached register C array.